Fix all clang-tidy warnings for green static analysis CI
Mechanical fixes across 22 files:
- myassert.h: parenthesize macro arguments (bugprone-macro-parentheses)
- LM headers: C-style casts to reinterpret_cast (cppcoreguidelines-pro-type-cstyle-cast, performance-no-int-to-ptr)
- PPMLanguageModel.h: const postfix operator++ (cert-dcl21-cpp)
- MandarinAlphMgr.h, RoutingAlphMgr.h, OneDimensionalFilter.cpp: static_cast to dynamic_cast (cppcoreguidelines-pro-type-static-cast-downcast)
- XmlSettingsStore.cpp, AlphIO.cpp: const ref for-range loops (performance-for-range-copy)
- AlphIO.cpp: remove dead stores (clang-analyzer-deadcode.DeadStores)
- CTWLanguageModel.cpp: fix memcmp/strcmp comparisons, merge branch-clone, remove dead store
- DictLanguageModel.cpp: add DivideZero guard, fix C-style casts
- DictLanguageModel.h: remove unused private field m_pAlphMap
- DasherModel.cpp, ExpansionPolicy.cpp: widen multiplication operands (bugprone-implicit-widening-of-multiplication-result)
- ConversionManager.h: merge identical branches (bugprone-branch-clone)
- Suppressed false positives with NOLINT: VirtualCall, parent-virtual-call, EnumCastOutOfRange, cert-msc30-c (rand())

+++ src/DasherCore/ExpansionPolicy.cpp
@@ -122,7 +122,7 @@ AmortizedPolicy::AmortizedPolicy(CDasherModel* pModel, unsigned int iNodeBudget,
 
 double AmortizedPolicy::pushNode(CDasherNode* node, int iMin, int iMax, bool bExpand, double dParentCost) {
     double dRes = BudgettingPolicy::pushNode(node, iMin, iMax, bExpand, dParentCost);
-    if (bExpand && sExpand.size() > 2 * m_iMaxExpands) trim();
+    if (bExpand && sExpand.size() > 2 * static_cast<size_t>(m_iMaxExpands)) trim();
     return dRes;
 }
